# frontend/openapi.json
{"openapi": "3.1.0", "info": {"title": "FastAPI", "version": "0.1.0"}, "paths": {"/admin/users/{username}/activate": {"patch": {"tags": ["admin_user_commands"], "summary": "Activate User", "operationId": "admin_user_commands_activate_user", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "username", "in": "path", "required": true, "schema": {"type": "string", "title": "Username"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/admin/users/{username}/deactivate": {"patch": {"tags": ["admin_user_commands"], "summary": "Deactivate User", "operationId": "admin_user_commands_deactivate_user", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "username", "in": "path", "required": true, "schema": {"type": "string", "title": "Username"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/admin/users/{username}/password": {"patch": {"tags": ["admin_user_commands"], "summary": "Change Password", "operationId": "admin_user_commands_change_password", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "username", "in": "path", "required": true, "schema": {"type": "string", "title": "Username"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ChangePasswordWithoutOldPasswordRequest"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/admin/users/{username}/email": {"patch": {"tags": ["admin_user_commands"], "summary": "Change Email", "operationId": "admin_user_commands_change_email", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "username", "in": "path", "required": true, "schema": {"type": "string", "title": "Username"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ChangeEmailRequest"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/auth/register": {"post": {"tags": ["auth"], "summary": "Register", "operationId": "auth_register", "requestBody": {"content": {"application/x-www-form-urlencoded": {"schema": {"$ref": "#/components/schemas/Body_register_auth_register_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/auth/login": {"post": {"tags": ["auth"], "summary": "Login", "operationId": "auth_login", "requestBody": {"content": {"application/x-www-form-urlencoded": {"schema": {"$ref": "#/components/schemas/Body_login_auth_login_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/LoginUserResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/auth/activate": {"post": {"tags": ["auth"], "summary": "Activate User", "operationId": "auth_activate_user", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ActivateUserWithCodeRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/auth/resend-activation": {"post": {"tags": ["auth"], "summary": "Resend Activation Code", "operationId": "auth_resend_activation_code", "requestBody": {"content": {"application/x-www-form-urlencoded": {"schema": {"$ref": "#/components/schemas/Body_resend_activation_code_auth_resend_activation_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/users/me": {"get": {"tags": ["user_queries"], "summary": "Get Me", "operationId": "user_queries_get_me", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserResponse"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/users/": {"get": {"tags": ["user_queries"], "summary": "Get Users", "operationId": "user_queries_get_users", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/PublicUserResponse"}, "type": "array", "title": "Response Get Users Users  Get"}}}}}}}, "/users/{username}": {"get": {"tags": ["user_queries"], "summary": "Get User", "operationId": "user_queries_get_user", "parameters": [{"name": "username", "in": "path", "required": true, "schema": {"type": "string", "title": "Username"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PublicUserResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/users/me/password": {"patch": {"tags": ["user_commands"], "summary": "Change Password", "operationId": "user_commands_change_password", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ChangePasswordByUserRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/users/me/email": {"patch": {"tags": ["user_commands"], "summary": "Change Email", "operationId": "user_commands_change_email", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ChangeEmailRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/wishlists/": {"get": {"tags": ["wishlist_queries"], "summary": "Get Current User Wishlists", "operationId": "wishlist_queries_get_current_user_wishlists", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/WishlistResponse"}, "type": "array", "title": "Response Get Current User Wishlists Wishlists  Get"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/wishlists/archived": {"get": {"tags": ["wishlist_queries"], "summary": "Get Archived Wishlists", "operationId": "wishlist_queries_get_archived_wishlists", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/WishlistResponse"}, "type": "array", "title": "Response Get Archived Wishlists Wishlists Archived Get"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/wishlists/{uuid}": {"get": {"tags": ["wishlist_queries"], "summary": "Get Wishlist", "operationId": "wishlist_queries_get_wishlist", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "uuid", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Uuid"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/WishlistResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/wishlists/user/{username}": {"get": {"tags": ["wishlist_queries"], "summary": "Get Wishlists By User", "operationId": "wishlist_queries_get_wishlists_by_user", "parameters": [{"name": "username", "in": "path", "required": true, "schema": {"type": "string", "title": "Username"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/WishlistResponse"}, "title": "Response Get Wishlists By User Wishlists User  Username  Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/wishlists/create": {"post": {"tags": ["wishlist_commands"], "summary": "Create Wishlist", "operationId": "wishlist_commands_create_wishlist", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateWishlistRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/wishlists/change-name/{wishlist_uuid}": {"post": {"tags": ["wishlist_commands"], "summary": "Change Wishlist Name", "operationId": "wishlist_commands_change_wishlist_name", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "wishlist_uuid", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Wishlist Uuid"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ChangeWishlistNameRequest"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/wishlists/archive/{wishlist_uuid}": {"post": {"tags": ["wishlist_commands"], "summary": "Archive Wishlist", "operationId": "wishlist_commands_archive_wishlist", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "wishlist_uuid", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Wishlist Uuid"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/wishlists/unarchive/{wishlist_uuid}": {"post": {"tags": ["wishlist_commands"], "summary": "Unarchive Wishlist", "operationId": "wishlist_commands_unarchive_wishlist", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "wishlist_uuid", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Wishlist Uuid"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/wishlists/set-visibility/{wishlist_uuid}": {"post": {"tags": ["wishlist_commands"], "summary": "Set Wishlist Visibility", "operationId": "wishlist_commands_set_wishlist_visibility", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "wishlist_uuid", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Wishlist Uuid"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ChangeWishlistVisibilityRequest"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/wishlists/add-item/{wishlist_uuid}": {"post": {"tags": ["wishlist_commands"], "summary": "Add Wishlist Item", "operationId": "wishlist_commands_add_wishlist_item", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "wishlist_uuid", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Wishlist Uuid"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AddWishlistItemRequest"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/wishlists/remove-item/{wishlist_uuid}": {"post": {"tags": ["wishlist_commands"], "summary": "Remove Wishlist Item", "operationId": "wishlist_commands_remove_wishlist_item", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "wishlist_uuid", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Wishlist Uuid"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RemoveWishlistItemRequest"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/wishlists/mark-item-as-purchased/{wishlist_uuid}": {"post": {"tags": ["wishlist_commands"], "summary": "Mark Item As Purchased", "operationId": "wishlist_commands_mark_item_as_purchased", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "wishlist_uuid", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Wishlist Uuid"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SetWishlistItemStatusRequest"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/wishlists/mark-item-as-not-purchased/{wishlist_uuid}": {"post": {"tags": ["wishlist_commands"], "summary": "Mark Item As Not Purchased", "operationId": "wishlist_commands_mark_item_as_not_purchased", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "wishlist_uuid", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Wishlist Uuid"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SetWishlistItemStatusRequest"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/health/live": {"get": {"tags": ["health"], "summary": "Liveness", "description": "Application is alive.", "operationId": "health_liveness", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/health/ready": {"get": {"tags": ["health"], "summary": "Readiness", "operationId": "health_readiness", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}}, "components": {"schemas": {"ActivateUserWithCodeRequest": {"properties": {"username": {"type": "string", "title": "Username"}, "code": {"type": "string", "pattern": "^\\d{8}$", "title": "Code"}}, "type": "object", "required": ["username", "code"], "title": "ActivateUserWithCodeRequest"}, "AddWishlistItemRequest": {"properties": {"name": {"type": "string", "title": "Name"}, "quantity": {"type": "integer", "title": "Quantity"}, "measurement_unit": {"$ref": "#/components/schemas/MeasurementUnit"}, "priority": {"$ref": "#/components/schemas/Priority"}}, "type": "object", "required": ["name", "quantity", "measurement_unit", "priority"], "title": "AddWishlistItemRequest"}, "Body_login_auth_login_post": {"properties": {"grant_type": {"anyOf": [{"type": "string", "pattern": "password"}, {"type": "null"}], "title": "Grant Type"}, "username": {"type": "string", "title": "Username"}, "password": {"type": "string", "title": "Password"}, "scope": {"type": "string", "title": "Scope", "default": ""}, "client_id": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Client Id"}, "client_secret": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Client Secret"}}, "type": "object", "required": ["username", "password"], "title": "Body_login_auth_login_post"}, "Body_register_auth_register_post": {"properties": {"username": {"type": "string", "title": "Username"}, "email": {"type": "string", "title": "Email"}, "password": {"type": "string", "title": "Password"}}, "type": "object", "required": ["username", "email", "password"], "title": "Body_register_auth_register_post"}, "Body_resend_activation_code_auth_resend_activation_post": {"properties": {"grant_type": {"anyOf": [{"type": "string", "pattern": "password"}, {"type": "null"}], "title": "Grant Type"}, "username": {"type": "string", "title": "Username"}, "password": {"type": "string", "title": "Password"}, "scope": {"type": "string", "title": "Scope", "default": ""}, "client_id": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Client Id"}, "client_secret": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Client Secret"}}, "type": "object", "required": ["username", "password"], "title": "Body_resend_activation_code_auth_resend_activation_post"}, "ChangeEmailRequest": {"properties": {"new_email": {"type": "string", "title": "New Email"}}, "type": "object", "required": ["new_email"], "title": "ChangeEmailRequest"}, "ChangePasswordByUserRequest": {"properties": {"old_password": {"type": "string", "title": "Old Password"}, "new_password": {"type": "string", "title": "New Password"}}, "type": "object", "required": ["old_password", "new_password"], "title": "ChangePasswordByUserRequest"}, "ChangePasswordWithoutOldPasswordRequest": {"properties": {"new_password": {"type": "string", "title": "New Password"}}, "type": "object", "required": ["new_password"], "title": "ChangePasswordWithoutOldPasswordRequest"}, "ChangeWishlistNameRequest": {"properties": {"new_name": {"type": "string", "title": "New Name"}}, "type": "object", "required": ["new_name"], "title": "ChangeWishlistNameRequest"}, "ChangeWishlistVisibilityRequest": {"properties": {"is_public": {"type": "boolean", "title": "Is Public"}}, "type": "object", "required": ["is_public"], "title": "ChangeWishlistVisibilityRequest"}, "CreateWishlistRequest": {"properties": {"wishlist_name": {"type": "string", "title": "Wishlist Name"}, "is_public": {"type": "boolean", "title": "Is Public"}}, "type": "object", "required": ["wishlist_name", "is_public"], "title": "CreateWishlistRequest"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "LoginUserResponse": {"properties": {"access_token": {"type": "string", "title": "Access Token"}, "token_type": {"type": "string", "title": "Token Type"}}, "type": "object", "required": ["access_token", "token_type"], "title": "LoginUserResponse"}, "MeasurementUnit": {"type": "string", "enum": ["piece", "meter", "kg."], "title": "MeasurementUnit"}, "Priority": {"type": "integer", "enum": [1, 2, 3], "title": "Priority"}, "PublicUserResponse": {"properties": {"username": {"type": "string", "title": "Username"}}, "type": "object", "required": ["username"], "title": "PublicUserResponse"}, "RemoveWishlistItemRequest": {"properties": {"item_uuid": {"type": "string", "format": "uuid", "title": "Item Uuid"}}, "type": "object", "required": ["item_uuid"], "title": "RemoveWishlistItemRequest"}, "SetWishlistItemStatusRequest": {"properties": {"item_uuid": {"type": "string", "format": "uuid", "title": "Item Uuid"}}, "type": "object", "required": ["item_uuid"], "title": "SetWishlistItemStatusRequest"}, "UserResponse": {"properties": {"username": {"type": "string", "title": "Username"}, "email": {"type": "string", "title": "Email"}, "is_active": {"type": "boolean", "title": "Is Active"}, "is_superuser": {"type": "boolean", "title": "Is Superuser"}}, "type": "object", "required": ["username", "email", "is_active", "is_superuser"], "title": "UserResponse"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}, "WishlistItemResponse": {"properties": {"uuid": {"type": "string", "format": "uuid", "title": "Uuid"}, "wishlist_uuid": {"type": "string", "format": "uuid", "title": "Wishlist Uuid"}, "name": {"type": "string", "title": "Name"}, "quantity": {"type": "integer", "title": "Quantity"}, "measurement_unit": {"$ref": "#/components/schemas/MeasurementUnit"}, "priority": {"$ref": "#/components/schemas/Priority"}, "is_purchased": {"type": "boolean", "title": "Is Purchased"}}, "type": "object", "required": ["uuid", "wishlist_uuid", "name", "quantity", "measurement_unit", "priority", "is_purchased"], "title": "WishlistItemResponse"}, "WishlistResponse": {"properties": {"uuid": {"type": "string", "format": "uuid", "title": "Uuid"}, "owner_username": {"type": "string", "title": "Owner Username"}, "name": {"type": "string", "title": "Name"}, "items": {"items": {"$ref": "#/components/schemas/WishlistItemResponse"}, "type": "array", "title": "Items"}, "is_archived": {"type": "boolean", "title": "Is Archived"}, "is_public": {"type": "boolean", "title": "Is Public"}, "created_at": {"type": "string", "title": "Created At"}}, "type": "object", "required": ["uuid", "owner_username", "name", "items", "is_archived", "is_public", "created_at"], "title": "WishlistResponse"}}, "securitySchemes": {"OAuth2PasswordBearer": {"type": "oauth2", "flows": {"password": {"scopes": {}, "tokenUrl": "/auth/login"}}}}}}